Absorb GitHub integration into facade
