+++ GEMINI.md
@@ -29,7 +29,7 @@ The architecture preserves SRP seams across the remaining workspace crates and a
 | `perfgate-render` | Workspace-only compatibility wrapper for `perfgate::presentation::render` |
 | `perfgate-export` | Workspace-only compatibility wrapper for `perfgate::presentation::export` |
 | `perfgate-sensor` | Workspace-only compatibility wrapper for `perfgate::presentation::sensor` |
-| `perfgate-github` | GitHub API and PR-comment integration |
+| `perfgate-github` | Workspace-only compatibility wrapper for `perfgate::integrations::github` |
 | `perfgate-domain::scaling` | Complexity and scaling analysis |
 | `perfgate-server` | Centralized Baseline Service API (REST/Axum) |
 | `perfgate-client` | Client library for Baseline Service interaction |
