+++ crates/perfgate-github/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "perfgate-github"
-description = "GitHub API client for perfgate PR comments"
+description = "Workspace-only compatibility wrapper for perfgate::integrations::github"
 version.workspace = true
 edition.workspace = true
 rust-version.workspace = true
@@ -10,20 +10,10 @@ homepage.workspace = true
 authors.workspace = true
 keywords = ["github", "pull-request", "performance", "comments"]
 categories = ["development-tools"]
+publish = false
 
 [dependencies]
-anyhow.workspace = true
-perfgate-app.workspace = true
-perfgate-types.workspace = true
-reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls-webpki-roots"] }
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-thiserror = "2"
-tracing = "0.1"
-
-[dev-dependencies]
-tokio = { version = "1", features = ["test-util", "macros", "rt-multi-thread"] }
-wiremock = "0.6"
+perfgate = { workspace = true, features = ["github"] }
 
 [lints]
 workspace = true
