+++ crates/perfgate-github/src/lib.rs
@@ -1,24 +1,5 @@
-//! # perfgate-github
+//! Compatibility wrapper for perfgate GitHub integration APIs.
 //!
-//! GitHub API client and comment rendering for perfgate PR comments.
-//!
-//! This crate provides:
-//!
-//! - A GitHub REST API client for creating, updating, and finding PR comments
-//! - Rich Markdown comment rendering with verdict badges, metric tables, trend indicators,
-//!   and blame attribution
-//! - Idempotent comment updates via a marker comment (`<!-- perfgate -->`)
-//! - Support for both GitHub Actions (GITHUB_TOKEN) and personal access tokens
+//! New code should import these APIs through `perfgate::integrations::github`.
 
-pub mod client;
-pub mod comment;
-pub mod error;
-pub mod types;
-
-pub use client::{COMMENT_MARKER, GitHubClient};
-pub use comment::{
-    CommentOptions, parse_github_repository, parse_pr_number_from_ref, render_comment,
-    render_comment_from_report,
-};
-pub use error::GitHubError;
-pub use types::GitHubComment;
+pub use perfgate::integrations::github::*;
