+++ crates/perfgate/Cargo.toml
@@ -25,13 +25,18 @@ serde_json.workspace = true
 thiserror.workspace = true
 time.workspace = true
 uuid.workspace = true
+reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls-webpki-roots"], optional = true }
+tracing = { version = "0.1", optional = true }
 
 [dev-dependencies]
 serde_json.workspace = true
 tempfile.workspace = true
+tokio = { version = "1", features = ["test-util", "macros", "rt-multi-thread"] }
+wiremock.workspace = true
 
 [features]
 default = []
+github = ["dep:reqwest", "dep:tracing"]
 
 [lints]
 workspace = true
