+++ crates/perfgate/src/integrations.rs
@@ -1,3 +1,5 @@
 //! Integrations for importing external benchmark data and CI platform output.
 
+#[cfg(feature = "github")]
+pub mod github;
 pub mod ingest;
