+++ crates/perfgate/src/integrations/github/comment.rs
@@ -3,7 +3,7 @@
 //! Produces rich Markdown that includes a verdict badge, metric comparison table,
 //! trend indicators, optional blame attribution, and collapsible raw data.
 
-use crate::client::COMMENT_MARKER;
+use super::client::COMMENT_MARKER;
 use perfgate_app::render::{
     direction_str, format_metric_with_statistic, format_pct, format_value, metric_status_icon,
     render_reason_line,
