+++ crates/perfgate/src/integrations/github/mod.rs
@@ -1,0 +1,22 @@
+//! GitHub API client and comment rendering for perfgate PR comments.
+//!
+//! This module provides:
+//!
+//! - A GitHub REST API client for creating, updating, and finding PR comments
+//! - Rich Markdown comment rendering with verdict badges, metric tables, trend indicators,
+//!   and blame attribution
+//! - Idempotent comment updates via a marker comment (`<!-- perfgate -->`)
+//! - Support for both GitHub Actions (GITHUB_TOKEN) and personal access tokens
+
+pub mod client;
+pub mod comment;
+pub mod error;
+pub mod types;
+
+pub use client::{COMMENT_MARKER, GitHubClient};
+pub use comment::{
+    CommentOptions, parse_github_repository, parse_pr_number_from_ref, render_comment,
+    render_comment_from_report,
+};
+pub use error::GitHubError;
+pub use types::GitHubComment;
