+++ docs/WORKSPACE.md
@@ -10,6 +10,7 @@ This file is automatically generated by `cargo run -p xtask -- docs-sync`.
 | `perfgate-export` | Workspace-only compatibility wrapper for perfgate::presentation::export | 90% |
 | `perfgate-render` | Workspace-only compatibility wrapper for perfgate::presentation::render | 90% |
 | `perfgate-sensor` | Workspace-only compatibility wrapper for perfgate::presentation::sensor | 90% |
+| `perfgate-github` | Workspace-only compatibility wrapper for perfgate::integrations::github | 90% |
 | `perfgate-paired` | Paired benchmarking statistics (A/B testing) | 100% |
 | `perfgate-fake` | Test utilities and fake implementations | 70% |
